+++ cli/cli/commands/enclave/add/add.go
@@ -78,7 +78,7 @@ func run(cmd *cobra.Command, args []string) error {
 
 	ctx := context.Background()
 
-	engineManager, err := engine_manager.NewEngineManager()
+	engineManager, err := engine_manager.NewEngineManager(ctx)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating an engine manager.")
 	}