+++ cli/cli/commands/engine/stop/stop.go
@@ -20,7 +20,7 @@ func run(cmd *cobra.Command, args []string) error {
 	ctx := context.Background()
 	logrus.Infof("Stopping Kurtosis engine...")
 
-	engineManager, err := engine_manager.NewEngineManager()
+	engineManager, err := engine_manager.NewEngineManager(ctx)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating an engine manager")
 	}