+++ cli/cli/commands/module/exec/exec.go
@@ -157,7 +157,7 @@ func run(cmd *cobra.Command, args []string) error {
 		)
 	}
 
-	engineManager, err := engine_manager.NewEngineManager()
+	engineManager, err := engine_manager.NewEngineManager(ctx)
 	if err != nil {
 		return stacktrace.Propagate(err, "An error occurred creating an engine manager.")
 	}